+++ internal/codegen/golang/vineschema/test_helper_test.go
@@ -1,4 +1,4 @@
-package schema
+package vineschema
 
 import (
 	"testing"