+++ internal/codegen/golang/vineschema/type_model.go
@@ -1,4 +1,4 @@
-package schema
+package vineschema
 
 type _MemberSchema struct {
 	Name             string